requirements.txt
python-docx>=0.8.11
feedgen>=0.9.0
requests>=2.28.0
python-dateutil>=2.8.0
lxml>=4.6.0
pytz>=2022.1